don't overload log2

+++ src/information.h
@@ -25,11 +25,11 @@ __attribute__((constructor))
     l2[2] = 1;
     
     for (int_fast32_t i = 2; i != FACT_MAX; ++i) {
-      ldfact[i] = ldfact[i - 2] + log2(i);
+      ldfact[i] = ldfact[i - 2] + log2(double(i));
     }
     
     for (int_fast32_t i = 3; i != DAY_MAX_LEAVES; ++i) {
-      l2[i] = log2(i);
+      l2[i] = log2(double(i));
       l2rooted[i] = ldfact[(i << 1) - 3];
       assert(l2unrooted[i] == ldfact[(i << 1) - 5]);
     }
